Cleaned header, add autogenerated docs, add some docs
