+++ src/blockedtensorfactory.cc
@@ -27,6 +27,7 @@
  */
 
 #include "blockedtensorfactory.h"
+#include "psi4/psi4-dec.h"
 #include "psi4/libmints/molecule.h"
 #include "psi4/libmints/wavefunction.h"
 #include "psi4/liboptions/liboptions.h"
