+++ src/ci_reference.cc
@@ -29,6 +29,7 @@
 #include "psi4/libmints/molecule.h"
 #include "psi4/libmints/wavefunction.h"
 #include "psi4/liboptions/liboptions.h"
+#include "psi4/libpsi4util/PsiOutStream.h"
 
 #include "ci_reference.h"
