+++ src/fci/fci_vector_h_diag.cc
@@ -30,6 +30,9 @@
 #include <numeric>
 #include <vector>
 
+#include "psi4/libpsi4util/PsiOutStream.h"
+#include "psi4/psi4-dec.h"
+
 #include "../helpers.h"
 #include "fci_vector.h"
