+++ src/forte_options.cc
@@ -1,8 +1,16 @@
 #include "forte_options.h"
 #include "helpers.h"
 
+#include "psi4/libpsi4util/PsiOutStream.h"
+
 namespace psi {
 namespace forte {
+
+std::string rst_bold(const std::string& s);
+
+std::string option_formatter(const std::string& type, const std::string& label,
+                             const std::string& default_value, const std::string& description,
+                             const std::string& allowed_values);
 
 void ForteOptions::add_bool(const std::string& label, bool value, const std::string& description) {
     bool_opts_.push_back(std::make_tuple(label, value, description));
@@ -59,5 +67,118 @@ void ForteOptions::add_psi4_options(Options& options) {
         options.add(std::get<0>(opt), new ArrayType());
     }
 }
+
+std::string ForteOptions::generate_documentation() const {
+    //    // For the bool, int, and double types store:
+    //    // ("label", default value, "description")
+    //    using bool_opt_t = std::tuple<std::string, bool, std::string>;
+    //    using int_opt_t = std::tuple<std::string, int, std::string>;
+    //    using double_opt_t = std::tuple<std::string, double, std::string>;
+
+    //    // For the string type stores:
+    //    // ("label", default value, "description",vector<"allowed values">)
+    //    using str_opt_t = std::tuple<std::string, std::string, std::string,
+    //    std::vector<std::string>>;
+
+    //    // For the array type stores:
+    //    // ("label", "description")
+    //    using array_opt_t = std::tuple<std::string, std::string>;
+
+    std::vector<std::pair<std::string, std::string>> option_docs_list;
+
+    for (const auto& opt : bool_opts_) {
+        const std::string& label = std::get<0>(opt);
+        const std::string& default_value = std::get<1>(opt) ? "True" : "False";
+        const std::string& description = std::get<2>(opt);
+        std::string option_text =
+            option_formatter("Boolean", label, default_value, description, "");
+        outfile->Printf("\n %s", label.c_str());
+        option_docs_list.push_back(std::make_pair(label, option_text));
+    }
+
+    for (const auto& opt : int_opts_) {
+        const std::string& label = std::get<0>(opt);
+        const std::string& default_value = std::to_string(std::get<1>(opt));
+        const std::string& description = std::get<2>(opt);
+        std::string option_text =
+            option_formatter("Integer", label, default_value, description, "");
+        outfile->Printf("\n %s", label.c_str());
+        option_docs_list.push_back(std::make_pair(label, option_text));
+    }
+
+    for (const auto& opt : double_opts_) {
+        const std::string& label = std::get<0>(opt);
+        const std::string& default_value = std::to_string(std::get<1>(opt));
+        const std::string& description = std::get<2>(opt);
+        std::string option_text = option_formatter("Double", label, default_value, description, "");
+        outfile->Printf("\n %s", label.c_str());
+        option_docs_list.push_back(std::make_pair(label, option_text));
+    }
+
+    for (const auto& opt : str_opts_) {
+        const std::string& label = std::get<0>(opt);
+        const std::string& default_value = std::get<1>(opt);
+        const std::string& description = std::get<2>(opt);
+        const std::string& allowed_values = to_string(std::get<3>(opt), ", ");
+        std::string option_text =
+            option_formatter("String", label, default_value, description, allowed_values);
+        outfile->Printf("\n %s", label.c_str());
+        option_docs_list.push_back(std::make_pair(label, option_text));
+    }
+
+    for (const auto& opt : array_opts_) {
+        const std::string& label = std::get<0>(opt);
+        const std::string& description = std::get<1>(opt);
+        std::string option_text = option_formatter("Array", label, "[]", description, "");
+        outfile->Printf("\n %s", label.c_str());
+        option_docs_list.push_back(std::make_pair(label, option_text));
+    }
+
+    std::sort(option_docs_list.begin(), option_docs_list.end());
+    std::vector<std::string> options_lines;
+
+    options_lines.push_back(".. _`sec:options`:\n");
+    options_lines.push_back("List of Forte options");
+    options_lines.push_back("=====================\n");
+    options_lines.push_back(".. sectionauthor:: Francesco A. Evangelista\n");
+    for (const auto& p : option_docs_list) {
+        options_lines.push_back(p.second);
+    }
+
+    return to_string(options_lines, "\n");
 }
+
+std::string rst_bold(const std::string& s) { return "**" + s + "**"; }
+
+//"label", default value, "description",vector<"allowed values">)
+std::string option_formatter(const std::string& type, const std::string& label,
+                             const std::string& default_value, const std::string& description,
+                             const std::string& allowed_values) {
+    std::string s;
+
+    s += rst_bold(label) + "\n\n";
+    s += description + "\n\n";
+    s += "* Type: " + type + "\n\n";
+    s += "* Default value: " + default_value + "\n\n";
+    if (allowed_values.size() > 0) {
+        s += "* Allowed values: " + allowed_values;
+    }
+
+    //**INT_TYPE**
+
+    // If one is going to use THREE-DSRG-MRPT2, this keyword needs to be set to either DF or
+    // CHOLESKY.  The default value for integrals is actually conventional, but the code won't work
+    // if conventional is used.
+
+    // INT_TYPE tells what type of integrals will be used in the calculation
+
+    //* Type: string
+
+    //* Possible Values: DF, CHOLESKY
+
+    //* Default: CONVENTIONAL
+    return s;
 }
+
+} // namespace forte
+} // namespace psi
