+++ src/integrals/custom_integrals.cc
@@ -30,6 +30,7 @@
 
 #include "psi4/libdpd/dpd.h"
 #include "psi4/libmints/mintshelper.h"
+#include "psi4/libmints/matrix.h"
 #include "psi4/libpsio/psio.hpp"
 #include "psi4/psi4-dec.h"
 #include "psi4/psifiles.h"
