+++ src/integrals/df_integrals.cc
@@ -31,7 +31,7 @@
 
 #include "psi4/libpsi4util/process.h"
 #include "psi4/libmints/basisset.h"
-#include "psi4/libmints/basisset.h"
+#include "psi4/libmints/matrix.h"
 #include "psi4/libmints/wavefunction.h"
 #include "psi4/libqt/qt.h"
 #include "psi4/lib3index/dfhelper.h"
@@ -45,6 +45,7 @@
 #endif
 
 #include "../blockedtensorfactory.h"
+#include "df_integrals.h"
 
 using namespace ambit;
 namespace psi {
@@ -150,6 +151,12 @@ ambit::Tensor DFIntegrals::three_integral_block(const std::vector<size_t>& A,
         value = three_integral(A[i[0]], p[i[1]], q[i[2]]);
     });
     return ReturnTensor;
+}
+
+ambit::Tensor DFIntegrals::three_integral_block_two_index(const std::vector<size_t>&, size_t,
+                                                          const std::vector<size_t>&) {
+    outfile->Printf("\n Oh no! this isn't here");
+    throw PSIEXCEPTION("INT_TYPE=DISKDF");
 }
 
 void DFIntegrals::set_tei(size_t, size_t, size_t, size_t, double, bool, bool) {
