+++ src/integrals/integrals.h
@@ -36,16 +36,17 @@
 #include <iostream>
 #include <vector>
 
+#include "psi4/libmints/typedefs.h"
 #include "ambit/blocked_tensor.h"
-#include "psi4/libmints/matrix.h"
-#include "psi4/liboptions/liboptions.h"
-#include "psi4/libtrans/integraltransform.h"
-#include "psi4/libpsi4util/PsiOutStream.h"
-#include "psi4/lib3index/dfhelper.h"
+//#include "psi4/libmints/matrix.h"
 
 namespace psi {
 
 class Tensor;
+class Options;
+class Matrix;
+class Wavefunction;
+class Dimension;
 
 namespace forte {
 
@@ -83,7 +84,7 @@ class ForteIntegrals {
      * @param restricted Select a restricted or unrestricted transformation
      * @param mo_space_info The MOSpaceInfo object
      */
-    ForteIntegrals(psi::Options& options, SharedWavefunction ref_wfn,
+    ForteIntegrals(psi::Options& options, std::shared_ptr<Wavefunction> ref_wfn,
                    IntegralSpinRestriction restricted, std::shared_ptr<MOSpaceInfo> mo_space_info);
 
     /// Destructor
@@ -99,10 +100,7 @@ class ForteIntegrals {
     void set_print(int print) { print_ = print; }
 
     /// Return the number of auxiliary functions
-    virtual size_t nthree() const {
-        throw PSIEXCEPTION("WRONG INT_TYPE");
-        return 1;
-    }
+    virtual size_t nthree() const = 0;
 
     /// Return the frozen core energy
     double frozen_core_energy() { return frozen_core_energy_; }
@@ -129,24 +127,10 @@ class ForteIntegrals {
     std::vector<double> get_fock_b() const { return fock_matrix_b_; }
 
     /// Set the alpha fock matrix
-    void set_fock_a(const std::vector<double>& fock_stl) {
-        size_t fock_size = fock_stl.size();
-        if (fock_size != ncmo_ * ncmo_) {
-            throw PSIEXCEPTION("Cannot fill in fock_matrix_a because the vector is out-of-range.");
-        } else {
-            fock_matrix_a_ = fock_stl;
-        }
-    }
+    void set_fock_a(const std::vector<double>& fock_stl);
 
     /// Set the beta fock matrix
-    void set_fock_b(const std::vector<double>& fock_stl) {
-        size_t fock_size = fock_stl.size();
-        if (fock_size != ncmo_ * ncmo_) {
-            throw PSIEXCEPTION("Cannot fill in fock_matrix_b because the vector is out-of-range.");
-        } else {
-            fock_matrix_b_ = fock_stl;
-        }
-    }
+    void set_fock_b(const std::vector<double>& fock_stl);
 
     /// The antisymmetrixed alpha-alpha two-electron integrals in physicist
     /// notation <pq||rs>
@@ -186,7 +170,8 @@ class ForteIntegrals {
                                                          const std::vector<size_t>& q) = 0;
 
     /// Make a Fock matrix computed with respect to a given determinant
-    virtual void make_fock_matrix(SharedMatrix gamma_a, SharedMatrix gamma_b) = 0;
+    virtual void make_fock_matrix(std::shared_ptr<Matrix> gamma_a,
+                                  std::shared_ptr<Matrix> gamma_b) = 0;
 
     /// Set the value of the scalar part of the Hamiltonian
     /// @param value the new value of the scalar part of the Hamiltonian
@@ -216,9 +201,9 @@ class ForteIntegrals {
     /// Return the type of integral used
     IntegralType integral_type() { return integral_type_; }
     /// Return the one-body symmetry integrals
-    SharedMatrix OneBody_symm() { return OneBody_symm_; }
+    std::shared_ptr<Matrix> OneBody_symm() { return OneBody_symm_; }
     /// Return the one-body AO integrals
-    SharedMatrix OneBodyAO() { return OneIntsAO_; }
+    std::shared_ptr<Matrix> OneBodyAO() { return OneIntsAO_; }
 
     virtual int ga_handle() { return 0; }
 
@@ -227,7 +212,7 @@ class ForteIntegrals {
 
     /// Obtain AO dipole integrals [X, Y, Z]
     /// Each direction is a SharedMatrix of dimension nmo * nmo
-    std::vector<SharedMatrix> AOdipole_ints() { return AOdipole_ints_; }
+    std::vector<std::shared_ptr<Matrix>> AOdipole_ints() { return AOdipole_ints_; }
 
     /**
      * Compute MO dipole integrals
@@ -236,8 +221,8 @@ class ForteIntegrals {
      * @return a vector of MOdipole ints in X, Y, Z order,
      *         each of which is a nmo by nmo SharedMatrix
      */
-    std::vector<SharedMatrix> compute_MOdipole_ints(const bool& alpha = true,
-                                                    const bool& resort = false);
+    std::vector<std::shared_ptr<Matrix>> compute_MOdipole_ints(const bool& alpha = true,
+                                                               const bool& resort = false);
 
   protected:
     // ==> Class data <==
@@ -324,7 +309,7 @@ class ForteIntegrals {
 
     /// The B tensor
     // std::shared_ptr<psi::Tensor> B_;
-    std::shared_ptr<DFHelper> df_;
+    // remove this from general class    std::shared_ptr<DFHelper> df_;
 
     /// The type of tensor that ambit uses -> CoreTensor
     ambit::TensorType tensor_type_ = ambit::CoreTensor;
@@ -333,16 +318,16 @@ class ForteIntegrals {
     /// Control printing of timings
     int print_;
     /// The One Electron Integrals (T + V) in SO Basis
-    SharedMatrix OneBody_symm_;
-    SharedMatrix OneIntsAO_;
+    std::shared_ptr<Matrix> OneBody_symm_;
+    std::shared_ptr<Matrix> OneIntsAO_;
 
     /// AO dipole integrals
-    std::vector<SharedMatrix> AOdipole_ints_;
+    std::vector<std::shared_ptr<Matrix>> AOdipole_ints_;
     /// Compute AO dipole integrals
     void build_AOdipole_ints();
     /// Compute MO dipole integrals
-    std::vector<SharedMatrix> MOdipole_ints_helper(SharedMatrix Cao, SharedVector epsilon,
-                                                   const bool& resort);
+    std::vector<std::shared_ptr<Matrix>>
+    MOdipole_ints_helper(std::shared_ptr<Matrix> Cao, SharedVector epsilon, const bool& resort);
 
     // ==> Class private functions <==
 
@@ -372,417 +357,6 @@ class ForteIntegrals {
     /// Remove the doubly occupied and virtual orbitals and resort the rest so
     /// that we are left only with ncmo = nmo - nfzc - nfzv
     virtual void resort_integrals_after_freezing() = 0;
-    virtual void resort_three(std::shared_ptr<Matrix>&, std::vector<size_t>& map) = 0;
-};
-
-/**
- * @brief The ConventionalIntegrals class is an interface to calculate the
- * conventional integrals
- * Assumes storage of all tei and stores in core.
- */
-class ConventionalIntegrals : public ForteIntegrals {
-  public:
-    /// Contructor of the class.  Calls std::shared_ptr<ForteIntegrals> ints
-    /// constructor
-    ConventionalIntegrals(psi::Options& options, SharedWavefunction ref_wfn,
-                          IntegralSpinRestriction restricted,
-                          std::shared_ptr<MOSpaceInfo> mo_space_info);
-    virtual ~ConventionalIntegrals();
-
-    /// Grabs the antisymmetriced TEI - assumes storage in aphy_tei_*
-    virtual double aptei_aa(size_t p, size_t q, size_t r, size_t s);
-    virtual double aptei_ab(size_t p, size_t q, size_t r, size_t s);
-    virtual double aptei_bb(size_t p, size_t q, size_t r, size_t s);
-
-    /// Grabs the antisymmetrized TEI - assumes storage of ambit tensor
-    virtual ambit::Tensor aptei_aa_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-    virtual ambit::Tensor aptei_ab_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-    virtual ambit::Tensor aptei_bb_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-
-    //    virtual double three_integral(size_t, size_t, size_t) {
-    //        outfile->Printf("\n Oh no!, you tried to grab a ThreeIntegral but this "
-    //                        "is not there!!");
-    //        throw PSIEXCEPTION("INT_TYPE=DF/CHOLESKY to use ThreeIntegral");
-    //    }
-    virtual ambit::Tensor three_integral_block(const std::vector<size_t>&,
-                                               const std::vector<size_t>&,
-                                               const std::vector<size_t>&) {
-        outfile->Printf("\n Oh no!, you tried to grab a ThreeIntegral but this "
-                        "is not there!!");
-        throw PSIEXCEPTION("INT_TYPE=DF/CHOLESKY to use ThreeIntegral");
-    }
-    virtual ambit::Tensor three_integral_block_two_index(const std::vector<size_t>&, size_t,
-                                                         const std::vector<size_t>&) {
-        outfile->Printf("\n Oh no! this isn't here");
-        throw PSIEXCEPTION("INT_TYPE=DISKDF");
-    }
-
-    virtual double** three_integral_pointer() {
-        outfile->Printf("\n Doh! There is no Three_integral here.  Use DF/CD");
-        throw PSIEXCEPTION("INT_TYPE=DF/CHOLESKY to use ThreeIntegral!");
-    }
-
-    virtual void make_fock_matrix(SharedMatrix gamma_a, SharedMatrix gamma_b);
-
-    virtual size_t nthree() const { throw PSIEXCEPTION("Wrong Int_Type"); }
-
-  private:
-    /// Transform the integrals
-    void transform_integrals();
-
-    void resort_four(std::vector<double>& tei, std::vector<size_t>& map);
-
-    virtual void gather_integrals();
-    virtual void resort_integrals_after_freezing();
-    virtual void resort_three(std::shared_ptr<Matrix>&, std::vector<size_t>&) {}
-    virtual void set_tei(size_t p, size_t q, size_t r, size_t s, double value, bool alpha1,
-                         bool alpha2);
-
-    /// An addressing function to retrieve the two-electron integrals
-    size_t aptei_index(size_t p, size_t q, size_t r, size_t s) {
-        return aptei_idx_ * aptei_idx_ * aptei_idx_ * p + aptei_idx_ * aptei_idx_ * q +
-               aptei_idx_ * r + s;
-    }
-
-    /// The IntegralTransform object used by this class
-    std::shared_ptr<IntegralTransform> integral_transform_;
-
-    /// Two-electron integrals stored as a vector
-    std::vector<double> aphys_tei_aa;
-    std::vector<double> aphys_tei_ab;
-    std::vector<double> aphys_tei_bb;
-};
-
-/// Classes written by Kevin Hannon
-///
-/**
- * @brief The CholeskyIntegrals:  An interface that computes the cholesky
- * integrals,
- * freezes the core, and creates fock matrices from determinant classes
- */
-class CholeskyIntegrals : public ForteIntegrals {
-  public:
-    CholeskyIntegrals(psi::Options& options, SharedWavefunction ref_wfn,
-                      IntegralSpinRestriction restricted,
-                      std::shared_ptr<MOSpaceInfo> mo_space_info);
-    virtual ~CholeskyIntegrals();
-    /// aptei_x will grab antisymmetriced integrals and creates DF/CD integrals
-    /// on the fly
-    virtual double aptei_aa(size_t p, size_t q, size_t r, size_t s);
-    virtual double aptei_ab(size_t p, size_t q, size_t r, size_t s);
-    virtual double aptei_bb(size_t p, size_t q, size_t r, size_t s);
-
-    /// Grabs the antisymmetrized TEI - assumes storage of ambit tensor
-    virtual ambit::Tensor aptei_aa_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-    virtual ambit::Tensor aptei_ab_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-    virtual ambit::Tensor aptei_bb_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-
-    double three_integral(size_t A, size_t p, size_t q) {
-        return ThreeIntegral_->get(p * aptei_idx_ + q, A);
-    }
-    virtual double** three_integral_pointer() { return ThreeIntegral_->pointer(); }
-    virtual ambit::Tensor three_integral_block(const std::vector<size_t>& A,
-                                               const std::vector<size_t>& p,
-                                               const std::vector<size_t>& q);
-    virtual ambit::Tensor three_integral_block_two_index(const std::vector<size_t>&, size_t,
-                                                         const std::vector<size_t>&) {
-        outfile->Printf("\n Oh no! this isn't here");
-        throw PSIEXCEPTION("INT_TYPE=DISKDF");
-    }
-    /// Do not use this if you are using CD/DF integrals
-    virtual void set_tei(size_t p, size_t q, size_t r, size_t s, double value, bool alpha1,
-                         bool alpha2);
-
-    virtual void make_fock_matrix(SharedMatrix gamma_a, SharedMatrix gamma_b);
-
-    virtual size_t nthree() const { return nthree_; }
-    SharedMatrix L_ao_;
-
-  private:
-    /// Computes Cholesky integrals
-    virtual void gather_integrals();
-    virtual void resort_three(std::shared_ptr<Matrix>& threeint, std::vector<size_t>& map);
-    virtual void resort_integrals_after_freezing();
-    void transform_integrals();
-    std::shared_ptr<Matrix> ThreeIntegral_;
-    size_t nthree_ = 0;
-};
-
-/**
- * @brief The DFIntegrals class - interface to get DF integrals, freeze core and
- * resort,
- * make fock matrices, and grab information about the space
- */
-class DFIntegrals : public ForteIntegrals {
-  public:
-    DFIntegrals(psi::Options& options, SharedWavefunction ref_wfn,
-                IntegralSpinRestriction restricted, std::shared_ptr<MOSpaceInfo> mo_space_info);
-    virtual double aptei_aa(size_t p, size_t q, size_t r, size_t s);
-    virtual double aptei_ab(size_t p, size_t q, size_t r, size_t s);
-    virtual double aptei_bb(size_t p, size_t q, size_t r, size_t s);
-
-    /// Reads the antisymmetrized alpha-alpha chunck and returns an
-    /// ambit::Tensor
-    /// Grabs the antisymmetrized TEI - assumes storage of ambit tensor
-    virtual ambit::Tensor aptei_aa_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-    virtual ambit::Tensor aptei_ab_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-    virtual ambit::Tensor aptei_bb_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-
-    double three_integral(size_t A, size_t p, size_t q) {
-        return ThreeIntegral_->get(p * aptei_idx_ + q, A);
-    }
-    virtual ambit::Tensor three_integral_block(const std::vector<size_t>& A,
-                                               const std::vector<size_t>& p,
-                                               const std::vector<size_t>& q);
-    virtual ambit::Tensor three_integral_block_two_index(const std::vector<size_t>&, size_t,
-                                                         const std::vector<size_t>&) {
-        outfile->Printf("\n Oh no! this isn't here");
-        throw PSIEXCEPTION("INT_TYPE=DISKDF");
-    }
-    virtual double** three_integral_pointer() { return ThreeIntegral_->pointer(); }
-    virtual void set_tei(size_t p, size_t q, size_t r, size_t s, double value, bool alpha1,
-                         bool alpha2);
-    virtual ~DFIntegrals();
-
-    virtual void make_fock_matrix(SharedMatrix gamma_a, SharedMatrix gamma_b);
-
-    virtual size_t nthree() const { return nthree_; }
-
-  private:
-    virtual void gather_integrals();
-    virtual void resort_three(std::shared_ptr<Matrix>& threeint, std::vector<size_t>& map);
-    virtual void resort_integrals_after_freezing();
-    virtual void resort_four(double*&, std::vector<size_t>&) {}
-
-    std::shared_ptr<Matrix> ThreeIntegral_;
-    size_t nthree_ = 0;
-};
-
-/// A DiskDFIntegrals class for avoiding the storage of the ThreeIntegral tensor
-/// Assumes that the DFIntegrals are stored in a binary file generated by
-/// DF_Helper
-/// Aptei_xy are extremely slow -> Try to use three_electron_block.  Much faster
-/// Reading individual elements is slow
-class DISKDFIntegrals : public ForteIntegrals {
-  public:
-    DISKDFIntegrals(psi::Options& options, SharedWavefunction ref_wfn,
-                    IntegralSpinRestriction restricted, std::shared_ptr<MOSpaceInfo> mo_space_info);
-
-    /// aptei_xy functions are slow.  try to use three_integral_block
-
-    virtual double aptei_aa(size_t p, size_t q, size_t r, size_t s);
-    virtual double aptei_ab(size_t p, size_t q, size_t r, size_t s);
-    virtual double aptei_bb(size_t p, size_t q, size_t r, size_t s);
-
-    /// Reads the antisymmetrized alpha-alpha chunck and returns an
-    /// ambit::Tensor
-    virtual ambit::Tensor aptei_aa_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-    virtual ambit::Tensor aptei_ab_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-    virtual ambit::Tensor aptei_bb_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s);
-
-    virtual double diag_aptei_aa(size_t p, size_t q);
-    virtual double diag_aptei_ab(size_t p, size_t q);
-    virtual double diag_aptei_bb(size_t p, size_t q);
-    //    virtual double three_integral(size_t A, size_t p, size_t q);
-    virtual double** three_integral_pointer() { return (ThreeIntegral_->pointer()); }
-    /// Read a block of the DFIntegrals and return an Ambit tensor of size A by
-    /// p by q
-    virtual ambit::Tensor three_integral_block(const std::vector<size_t>& A,
-                                               const std::vector<size_t>& p,
-                                               const std::vector<size_t>& q);
-    /// return ambit tensor of size A by q
-    virtual ambit::Tensor three_integral_block_two_index(const std::vector<size_t>& A, size_t p,
-                                                         const std::vector<size_t>& q);
-
-    virtual void set_tei(size_t p, size_t q, size_t r, size_t s, double value, bool alpha1,
-                         bool alpha2);
-    virtual ~DISKDFIntegrals();
-
-    virtual void make_fock_matrix(SharedMatrix gamma_a, SharedMatrix gamma_b);
-
-    /// Make a Fock matrix computed with respect to a given determinant
-    virtual size_t nthree() const { return nthree_; }
-
-  private:
-    virtual void gather_integrals();
-    virtual void resort_three(std::shared_ptr<Matrix>& threeint, std::vector<size_t>& map);
-    virtual void resort_integrals_after_freezing();
-
-    std::shared_ptr<Matrix> ThreeIntegral_;
-    size_t nthree_ = 0;
-};
-
-#ifdef HAVE_GA
-class DistDFIntegrals : public ForteIntegrals {
-  public:
-    DistDFIntegrals(psi::Options& options, SharedWavefunction ref_wfn,
-                    IntegralSpinRestriction restricted, std::shared_ptr<MOSpaceInfo> mo_space_info);
-
-    virtual void retransform_integrals();
-    /// aptei_xy functions are slow.  try to use three_integral_block
-
-    virtual double aptei_aa(size_t /*p*/, size_t /*q*/, size_t /*r*/, size_t /*s*/) {}
-    virtual double aptei_ab(size_t /*p*/, size_t /*q*/, size_t /*r*/, size_t /*s*/) {}
-    virtual double aptei_bb(size_t /*p*/, size_t /*q*/, size_t /*r*/, size_t /*s*/) {}
-
-    /// Reads the antisymmetrized alpha-alpha chunck and returns an
-    /// ambit::Tensor
-    virtual ambit::Tensor aptei_aa_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s) {}
-    virtual ambit::Tensor aptei_ab_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s) {}
-    virtual ambit::Tensor aptei_bb_block(const std::vector<size_t>& p, const std::vector<size_t>& q,
-                                         const std::vector<size_t>& r,
-                                         const std::vector<size_t>& s) {}
-
-    virtual double diag_aptei_aa(size_t, size_t) {}
-    virtual double diag_aptei_ab(size_t, size_t) {}
-    virtual double diag_aptei_bb(size_t, size_t) {}
-    virtual double three_integral(size_t, size_t, size_t) {}
-    virtual double** three_integral_pointer() {
-        throw PSIEXCEPTION("Integrals are distributed.  Pointer does not exist");
-    }
-    /// Read a block of the DFIntegrals and return an Ambit tensor of size A by
-    /// p by q
-    virtual ambit::Tensor three_integral_block(const std::vector<size_t>& /*A*/,
-                                               const std::vector<size_t>& /*p*/,
-                                               const std::vector<size_t>& /*q*/);
-    /// return ambit tensor of size A by q
-    virtual ambit::Tensor three_integral_block_two_index(const std::vector<size_t>& /*A*/,
-                                                         size_t /*p*/,
-                                                         const std::vector<size_t>& /*q*/) {}
-
-    virtual void set_tei(size_t, size_t, size_t, size_t, double, bool, bool) {
-        outfile->Printf("DistributedDF will not work with set_tei");
-        throw PSIEXCEPTION("DistDF can not use set_tei");
-    }
-    virtual ~DistDFIntegrals();
-
-    virtual void make_fock_matrix(SharedMatrix /*gamma_a*/, SharedMatrix /*gamma_b*/) {}
-
-    /// Make a Fock matrix computed with respect to a given determinant
-    virtual size_t nthree() const { return nthree_; }
-    virtual int ga_handle() { return DistDF_ga_; }
-
-  private:
-    virtual void gather_integrals();
-    virtual void resort_three(std::shared_ptr<Matrix>& /*threeint*/, std::vector<size_t>& /*map*/) {
-    }
-    virtual void resort_integrals_after_freezing() {}
-
-    /// This is the handle for GA
-    int DistDF_ga_;
-
-    std::shared_ptr<Matrix> ThreeIntegral_;
-    size_t nthree_ = 0;
-    /// Assuming integrals are stored on disk
-    /// Reads the block of integrals present for each process
-    ambit::Tensor read_integral_chunk(std::shared_ptr<Tensor>& B, std::vector<int>& lo,
-                                      std::vector<int>& hi);
-    /// Distributes tensor according to naux dimension
-    void create_dist_df();
-    void test_distributed_integrals();
-};
-#endif
-
-/// This class is used if the user wants to generate their own integrals for
-/// their method.
-/// This would be very useful for CI based methods (the integrals class is
-/// wasteful and dumb for this area)
-/// Also, I am putting this here if I(Kevin) ever get around to implementing
-/// AO-DSRG-MRPT2
-class OwnIntegrals : public ForteIntegrals {
-  public:
-    OwnIntegrals(psi::Options& options, SharedWavefunction ref_wfn,
-                 IntegralSpinRestriction restricted, std::shared_ptr<MOSpaceInfo> mo_space_info);
-
-    virtual void retransform_integrals() {}
-    /// aptei_xy functions are slow.  try to use three_integral_block
-
-    virtual double aptei_aa(size_t /*p*/, size_t /*q*/, size_t /*r*/, size_t /*s*/) { return 0.0; }
-    virtual double aptei_ab(size_t /*p*/, size_t /*q*/, size_t /*r*/, size_t /*s*/) { return 0.0; }
-    virtual double aptei_bb(size_t /*p*/, size_t /*q*/, size_t /*r*/, size_t /*s*/) { return 0.0; }
-
-    /// Reads the antisymmetrized alpha-alpha chunck and returns an
-    /// ambit::Tensor
-    virtual ambit::Tensor aptei_aa_block(const std::vector<size_t>& /*p*/,
-                                         const std::vector<size_t>& /*q*/,
-                                         const std::vector<size_t>& /*r*/,
-                                         const std::vector<size_t>& /*s*/) {
-        return blank_tensor_;
-    }
-    virtual ambit::Tensor aptei_ab_block(const std::vector<size_t>& /*p*/,
-                                         const std::vector<size_t>& /*q*/,
-                                         const std::vector<size_t>& /*r*/,
-                                         const std::vector<size_t>& /*s*/) {
-        return blank_tensor_;
-    }
-    virtual ambit::Tensor aptei_bb_block(const std::vector<size_t>& /*p*/,
-                                         const std::vector<size_t>& /*q*/,
-                                         const std::vector<size_t>& /*r*/,
-                                         const std::vector<size_t>& /*s*/) {
-        return blank_tensor_;
-    }
-
-    virtual double diag_aptei_aa(size_t, size_t) { return 0.0; }
-    virtual double diag_aptei_ab(size_t, size_t) { return 0.0; }
-    virtual double diag_aptei_bb(size_t, size_t) { return 0.0; }
-    virtual double three_integral(size_t, size_t, size_t) { return 0.0; }
-    virtual double** three_integral_pointer() {
-        throw PSIEXCEPTION("Integrals are distributed.  Pointer does not exist");
-    }
-    /// Read a block of the DFIntegrals and return an Ambit tensor of size A by
-    /// p by q
-    virtual ambit::Tensor three_integral_block(const std::vector<size_t>& /*A*/,
-                                               const std::vector<size_t>& /*p*/,
-                                               const std::vector<size_t>& /*q*/) {
-        return blank_tensor_;
-    }
-    /// return ambit tensor of size A by q
-    virtual ambit::Tensor three_integral_block_two_index(const std::vector<size_t>& /*A*/,
-                                                         size_t /*p*/,
-                                                         const std::vector<size_t>& /*q*/) {
-        return blank_tensor_;
-    }
-
-    virtual void set_tei(size_t, size_t, size_t, size_t, double, bool, bool) {}
-    virtual ~OwnIntegrals();
-
-    virtual void make_fock_matrix(SharedMatrix /*gamma_a*/, SharedMatrix /*gamma_b*/) {}
-    virtual size_t nthree() const { return 1; }
-
-  private:
-    virtual void gather_integrals() {}
-    virtual void resort_three(std::shared_ptr<Matrix>& /*threeint*/, std::vector<size_t>& /*map*/) {
-    }
-    virtual void resort_integrals_after_freezing() {}
-    ambit::Tensor blank_tensor_;
 };
 
 } // namespace forte
