+++ src/integrals/own_integrals.cc
@@ -31,8 +31,12 @@
 
 #include "psi4/libmints/basisset.h"
 #include "psi4/libqt/qt.h"
+#include "psi4/libpsi4util/PsiOutStream.h"
+#include "psi4/libpsi4util/exception.h"
 
 #include "../blockedtensorfactory.h"
+
+#include "own_integrals.h"
 
 namespace psi {
 namespace forte {
