+++ src/mrdsrg-spin-integrated/mrdsrg_commutator.cc
@@ -30,6 +30,7 @@
 #include <map>
 #include <vector>
 
+#include "psi4/libpsi4util/PsiOutStream.h"
 #include "../helpers.h"
 #include "../mini-boost/boost/format.hpp"
 #include "mrdsrg.h"
