+++ src/mrdsrg-spin-integrated/mrdsrg_pt.cc
@@ -33,6 +33,7 @@
 #include <vector>
 
 #include "psi4/libdiis/diismanager.h"
+#include "psi4/libpsi4util/PsiOutStream.h"
 
 #include "../helpers.h"
 #include "../mini-boost/boost/format.hpp"
