+++ src/orbital-helper/unpaired_density.cc
@@ -30,6 +30,8 @@
 #include "psi4/libpsio/psio.h"
 #include "psi4/libpsio/psio.hpp"
 
+#include "psi4/libpsi4util/PsiOutStream.h"
+
 #include "psi4/libmints/local.h"
 #include "psi4/libmints/matrix.h"
 #include "psi4/libmints/vector.h"
