+++ src/semi_canonicalize.cc
@@ -36,6 +36,8 @@
 #include "psi4/libmints/matrix.h"
 #include "psi4/libmints/vector.h"
 
+#include "psi4/libpsi4util/PsiOutStream.h"
+
 #include "../blockedtensorfactory.h"
 #include "helpers.h"
 #include "semi_canonicalize.h"
