+++ src/sparse_ci/stl_bitset_string.cc
@@ -28,6 +28,7 @@
 
 #include "psi4/libmints/matrix.h"
 #include "psi4/psi4-dec.h"
+#include "psi4/libpsi4util/PsiOutStream.h"
 
 #include "stl_bitset_string.h"
