+++ src/sparse_ci/ui64_determinant_functions.cc
@@ -30,6 +30,7 @@
 #include <algorithm>
 #include <unordered_map>
 
+#include "psi4/libmints/dimension.h"
 #include "psi4/libpsi4util/PsiOutStream.h"
 #include "psi4/psi4-dec.h"
