+++ src/tensorsrg.cc
@@ -35,6 +35,9 @@
 #include "psi4/libmints/molecule.h"
 #include "psi4/libmints/wavefunction.h"
 
+#include "psi4/libpsi4util/PsiOutStream.h"
+#include "psi4/psi4-dec.h"
+
 #include "helpers.h"
 #include "tensorsrg.h"
