+++ src/tensorsrg_hermitian_commutator.cc
@@ -28,6 +28,9 @@
 
 #include <cmath>
 
+#include "psi4/libpsi4util/PsiOutStream.h"
+#include "psi4/psi4-dec.h"
+
 #include "helpers.h"
 #include "tensorsrg.h"
