+++ src/tensorsrg_srg.cc
@@ -28,6 +28,9 @@
 
 #include <cmath>
 
+#include "psi4/libpsi4util/PsiOutStream.h"
+#include "psi4/psi4-dec.h"
+
 #include "mini-boost/boost/numeric/odeint.hpp"
 
 #include "tensorsrg.h"
